+++ bssopenapi-20230930/src/Models/DescribeCouponShrinkRequest.php
@@ -59,6 +59,11 @@ class DescribeCouponShrinkRequest extends Model
     public $expireStartDate;
 
     /**
+     * @var bool
+     */
+    public $includeShare;
+
+    /**
      * @var int
      */
     public $maxResults;
@@ -81,6 +86,11 @@ class DescribeCouponShrinkRequest extends Model
     /**
      * @var string
      */
+    public $productCode;
+
+    /**
+     * @var string
+     */
     public $status;
     protected $_name = [
         'couponId' => 'CouponId',
@@ -93,10 +103,12 @@ class DescribeCouponShrinkRequest extends Model
         'effectiveStartTime' => 'EffectiveStartTime',
         'expireEndDate' => 'ExpireEndDate',
         'expireStartDate' => 'ExpireStartDate',
+        'includeShare' => 'IncludeShare',
         'maxResults' => 'MaxResults',
         'nbid' => 'Nbid',
         'nextToken' => 'NextToken',
         'pageSize' => 'PageSize',
+        'productCode' => 'ProductCode',
         'status' => 'Status',
     ];
 
@@ -148,6 +160,10 @@ class DescribeCouponShrinkRequest extends Model
             $res['ExpireStartDate'] = $this->expireStartDate;
         }
 
+        if (null !== $this->includeShare) {
+            $res['IncludeShare'] = $this->includeShare;
+        }
+
         if (null !== $this->maxResults) {
             $res['MaxResults'] = $this->maxResults;
         }
@@ -162,6 +178,10 @@ class DescribeCouponShrinkRequest extends Model
 
         if (null !== $this->pageSize) {
             $res['PageSize'] = $this->pageSize;
+        }
+
+        if (null !== $this->productCode) {
+            $res['ProductCode'] = $this->productCode;
         }
 
         if (null !== $this->status) {
@@ -219,6 +239,10 @@ class DescribeCouponShrinkRequest extends Model
             $model->expireStartDate = $map['ExpireStartDate'];
         }
 
+        if (isset($map['IncludeShare'])) {
+            $model->includeShare = $map['IncludeShare'];
+        }
+
         if (isset($map['MaxResults'])) {
             $model->maxResults = $map['MaxResults'];
         }
@@ -233,6 +257,10 @@ class DescribeCouponShrinkRequest extends Model
 
         if (isset($map['PageSize'])) {
             $model->pageSize = $map['PageSize'];
+        }
+
+        if (isset($map['ProductCode'])) {
+            $model->productCode = $map['ProductCode'];
         }
 
         if (isset($map['Status'])) {