+++ bssopenapi-20230930/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-05-07 Version: 1.7.5
+- Update API DescribeCoupon: add request parameters IncludeShare.
+- Update API DescribeCoupon: add request parameters ProductCode.
+
+
 2026-03-30 Version: 1.7.4
 - Update API DescribeCoupon: add request parameters CouponTemplateIdList.